Fix generated code
As good as the generator for the project was, there were a few edges
that still needed clean-up. For one, a bunch of haml files had their
extensions including the html which is totally unneeded. For another,
many views were still in ERB. There were a bunch of lines in various
files that either used old-style hash syntax or had commented code, so
those were also replaced.

Finally, there was one minor UI tweak to give error messages a top
margin so they weren't abutting the nav.